fix: Ensure that objects with relations cannot be deleted KAN-190 (#142)
* add logic and exception handling if resource is in use

* Improve Coverage
KAN-190

* Add integration test
KAN-190

* Add delete restriction
KAN-190

* Remove unused test
KAN-190

* fix: frontend tests

* Remove merge conflicts
KAN-190

---------

+++ backend/src/main/kotlin/ch/zhaw/pm4/simonsays/service/EventService.kt
@@ -4,6 +4,7 @@ import ch.zhaw.pm4.simonsays.api.mapper.EventMapper
 import ch.zhaw.pm4.simonsays.api.types.EventCreateUpdateDTO
 import ch.zhaw.pm4.simonsays.api.types.EventDTO
 import ch.zhaw.pm4.simonsays.entity.Event
+import ch.zhaw.pm4.simonsays.exception.ResourceInUseException
 import ch.zhaw.pm4.simonsays.exception.ResourceNotFoundException
 import ch.zhaw.pm4.simonsays.repository.EventRepository
 import org.springframework.stereotype.Service
@@ -48,6 +49,14 @@ class EventService(
 
     fun deleteEvent(eventId: Long) {
         val event = getEventEntity(eventId)
+        if(
+                !event.menuItems.isNullOrEmpty() ||
+                !event.stations.isNullOrEmpty() ||
+                !event.ingredients.isNullOrEmpty() ||
+                !event.menus.isNullOrEmpty()
+        ) {
+            throw ResourceInUseException("Event is used and cannot be deleted")
+        }
         eventRepository.delete(event)
     }
